新年更新

+++ zb_system/function/c_system_common.php
@@ -7,7 +7,7 @@
  */
 
 /**
- * 得到请求方法(未必会准确的，比如SERVER没能某项，或是端口改过的)
+ * 得到请求方法(未必会准确的，比如SERVER没有某项，或是端口改过的)
  * @param $array
  * @return $string
  */
@@ -37,16 +37,20 @@ function GetWebServer() {
         return SERVER_UNKNOWN;
     }
     $webServer = strtolower($_SERVER['SERVER_SOFTWARE']);
-    if (strpos($webServer, 'apache')) {
+    if (strpos($webServer, 'apache') !== false) {
         return SERVER_APACHE;
-    } elseif (strpos($webServer, 'microsoft-iis')) {
+    } elseif (strpos($webServer, 'microsoft-iis') !== false) {
         return SERVER_IIS;
-    } elseif (strpos($webServer, 'nginx')) {
+    } elseif (strpos($webServer, 'nginx') !== false) {
         return SERVER_NGINX;
-    } elseif (strpos($webServer, 'lighttpd')) {
+    } elseif (strpos($webServer, 'lighttpd') !== false) {
         return SERVER_LIGHTTPD;
-    } elseif (strpos($webServer, 'kangle')) {
+    } elseif (strpos($webServer, 'kangle') !== false) {
         return SERVER_KANGLE;
+    } elseif (strpos($webServer, 'caddy') !== false) {
+        return SERVER_CADDY;
+    } elseif (strpos($webServer, 'development server') !== false) {
+        return SERVER_BUILTIN;
     } else {
         return SERVER_UNKNOWN;
     }
@@ -87,6 +91,16 @@ function GetPHPEngine() {
 }
 
 /**
+ * 获取PHP Version
+ * @return string
+ */
+function GetPHPVersion() {
+    $p = phpversion();
+    if (strpos($p, '-') !== false) $p = substr($p, 0, strpos($p, '-'));
+    return $p;
+}
+
+/**
  * 自动加载类文件
  * @api Filter_Plugin_Autoload
  * @param string $classname 类名
@@ -94,9 +108,8 @@ function GetPHPEngine() {
  */
 function AutoloadClass($classname) {
     foreach ($GLOBALS['hooks']['Filter_Plugin_Autoload'] as $fpname => &$fpsignal) {
-        $fpsignal = PLUGIN_EXITSIGNAL_NONE;
         $fpreturn = $fpname($classname);
-        if ($fpsignal == PLUGIN_EXITSIGNAL_RETURN) {return $fpreturn;}
+        if ($fpsignal == PLUGIN_EXITSIGNAL_RETURN) {$fpsignal = PLUGIN_EXITSIGNAL_NONE;return $fpreturn;}
     }
     if (is_readable($f = ZBP_PATH . 'zb_system/function/lib/' . strtolower($classname) . '.php')) {
         require $f;
@@ -112,9 +125,8 @@ function AutoloadClass($classname) {
 function Logs($s, $iserror = false) {
     global $zbp;
     foreach ($GLOBALS['hooks']['Filter_Plugin_Logs'] as $fpname => &$fpsignal) {
-        $fpsignal = PLUGIN_EXITSIGNAL_NONE;
         $fpreturn = $fpname($s, $iserror);
-        if ($fpsignal == PLUGIN_EXITSIGNAL_RETURN) {return $fpreturn;}
+        if ($fpsignal == PLUGIN_EXITSIGNAL_RETURN) {$fpsignal = PLUGIN_EXITSIGNAL_NONE;return $fpreturn;}
     }
     if ($zbp->guid) {
         if ($iserror) {
@@ -185,12 +197,11 @@ function GetEnvironment() {
     if ($ajax) {
         $ajax = substr(get_class($ajax), 7);
     }
-
     $system_environment = PHP_OS . '; ' .
     GetValueInArray(
         explode(' ', str_replace(array('Microsoft-', '/'), array('', ''), GetVars('SERVER_SOFTWARE', 'SERVER'))), 0
     ) . '; ' .
-    'PHP ' . phpversion() . (IS_X64 ? ' x64' : '') . '; ' .
+    'PHP ' . GetPHPVersion() . (IS_X64 ? ' x64' : '') . '; ' .
     $zbp->option['ZC_DATABASE_TYPE'] . '; ' . $ajax;
 
     return $system_environment;
@@ -861,7 +872,7 @@ function JsonError($errorCode, $errorString, $data) {
             'timestamp' => time(),
         ),
     );
-    ob_clean();
+    @ob_clean();
     echo json_encode($result);
     if ($errorCode != 0) {
         exit;
@@ -918,6 +929,8 @@ function ScriptError($faultString) {
 function CheckRegExp($source, $para) {
     if (strpos($para, '[username]') !== false) {
         $para = "/^[\.\_A-Za-z0-9·@\x{4e00}-\x{9fa5}]+$/u";
+    } elseif (strpos($para, '[nickname]') !== false) {
+        $para = '/([^\x{01}-\x{1F}\x{80}-\x{FF}\/:\\~&%;@\'"?<>|#$\*}{,\+=\[\]\(\)\{\}\t\r\n\p{C}])/u';
     } elseif (strpos($para, '[password]') !== false) {
         $para = "/^[A-Za-z0-9`~!@#\$%\^&\*\-_\?]+$/u";
     } elseif (strpos($para, '[email]') !== false) {
@@ -1308,6 +1321,34 @@ function utf84mb_convertToUTF8($matches) {
 }
 
 
+//$args = 2...x
+function VerifyWebToken($wt,$wt_id){
+    $time = substr($wt,32);
+    $wt = substr($wt,0,32);
+    $args = array();
+    for ($i = 2; $i < func_num_args() ; $i++) { 
+        $args[] = func_get_arg($i);
+    }
+    $sha = md5( hash("sha256", $time . $wt_id) . hash("sha256", implode($args) . $time) );
+    if ($wt === $sha){
+        if ($time > time()){
+            return true;
+        }
+    }
+
+    return false;
+}
+//$time : expired second
+function CreateWebToken($wt_id,$time){
+    $time = (int)$time;
+    $args = array();
+    for ($i = 2; $i < func_num_args() ; $i++) { 
+        $args[] = func_get_arg($i);
+    }
+    return md5( hash("sha256", $time . $wt_id) . hash("sha256", implode($args) . $time) ) . $time;
+}
+
+
 
 /**
  * 处理PHP版本兼容代码
@@ -1328,19 +1369,34 @@ if (!function_exists('hex2bin')) {
 if (!function_exists('rrmdir')) {
     function rrmdir($dir) {
         if (is_dir($dir)) {
-            $objects = scandir($dir);
-            foreach ($objects as $object) {
-                if ($object != '.' && $object != '..') {
-                    if (filetype($dir . '/' . $object) == 'dir') {
-                        rrmdir($dir . '/' . $object);
-                    } else {
-                        unlink($dir . '/' . $object);
+            if (function_exists('scandir')) {
+                $objects = scandir($dir);
+                foreach ($objects as $object) {
+                    if ($object != '.' && $object != '..') {
+                        if (filetype($dir . '/' . $object) == 'dir') {
+                            rrmdir($dir . '/' . $object);
+                        } else {
+                            unlink($dir . '/' . $object);
+                        }
                     }
-
+                }
+                reset($objects);
+                rmdir($dir);
+            } else {
+                if ($handle = opendir($dir)) {
+                    while (false !== ($file = readdir($handle))) {
+                        if ($file != "." && $file != "..") {
+                            if (is_dir(rtrim(rtrim($dir, '/'), '\\') . '/' . $file)) {
+                                rrmdir(rtrim(rtrim($dir, '/'), '\\') . '/' . $file);
+                            } else {
+                                unlink(rtrim(rtrim($dir, '/'), '\\') . '/' . $file);
+                            }
+                        }
+                    }
+                    closedir($handle);
+                    rmdir($dir);
                 }
             }
-            reset($objects);
-            rmdir($dir);
         }
     }
 }
@@ -1500,20 +1556,20 @@ if (!function_exists('http_build_url')) {
 if (!function_exists('gzdecode')) {
  function gzdecode($data) {
    $len = strlen($data);
-   if ($len < 18 || strcmp(substr($data,0,2),"\x1f\x8b")) {
+   if ($len < 18 || strcmp(substr($data, 0, 2), "\x1f\x8b")) {
      return null;  // Not GZIP format (See RFC 1952)
    }
-   $method = ord(substr($data,2,1));  // Compression method
-   $flags  = ord(substr($data,3,1));  // Flags
+   $method = ord(substr($data, 2, 1));  // Compression method
+   $flags  = ord(substr($data, 3, 1));  // Flags
    if ($flags & 31 != $flags) {
      // Reserved bits are set -- NOT ALLOWED by RFC 1952
      return null;
    }
    // NOTE: $mtime may be negative (PHP integer limitations)
-   $mtime = unpack("V", substr($data,4,4));
+   $mtime = unpack("V", substr($data, 4, 4));
    $mtime = $mtime[1];
-   $xfl   = substr($data,8,1);
-   $os    = substr($data,8,1);
+   $xfl   = substr($data, 8, 1);
+   $os    = substr($data, 8, 1);
    $headerlen = 10;
    $extralen  = 0;
    $extra     = "";
@@ -1522,12 +1578,12 @@ if (!function_exists('gzdecode')) {
      if ($len - $headerlen - 2 < 8) {
        return false;    // Invalid format
      }
-     $extralen = unpack("v",substr($data,8,2));
+     $extralen = unpack("v", substr($data, 8, 2));
      $extralen = $extralen[1];
      if ($len - $headerlen - 2 - $extralen < 8) {
        return false;    // Invalid format
      }
-     $extra = substr($data,10,$extralen);
+     $extra = substr($data, 10, $extralen);
      $headerlen += 2 + $extralen;
    }
 
@@ -1538,11 +1594,11 @@ if (!function_exists('gzdecode')) {
      if ($len - $headerlen - 1 < 8) {
        return false;    // Invalid format
      }
-     $filenamelen = strpos(substr($data,8+$extralen),chr(0));
+     $filenamelen = strpos(substr($data, 8 + $extralen), chr(0));
      if ($filenamelen === false || $len - $headerlen - $filenamelen - 1 < 8) {
        return false;    // Invalid format
      }
-     $filename = substr($data,$headerlen,$filenamelen);
+     $filename = substr($data, $headerlen, $filenamelen);
      $headerlen += $filenamelen + 1;
    }
 
@@ -1553,11 +1609,11 @@ if (!function_exists('gzdecode')) {
      if ($len - $headerlen - 1 < 8) {
        return false;    // Invalid format
      }
-     $commentlen = strpos(substr($data,8+$extralen+$filenamelen),chr(0));
+     $commentlen = strpos(substr($data, 8 + $extralen + $filenamelen), chr(0));
      if ($commentlen === false || $len - $headerlen - $commentlen - 1 < 8) {
        return false;    // Invalid header format
      }
-     $comment = substr($data,$headerlen,$commentlen);
+     $comment = substr($data, $headerlen, $commentlen);
      $headerlen += $commentlen + 1;
    }
 
@@ -1567,8 +1623,8 @@ if (!function_exists('gzdecode')) {
      if ($len - $headerlen - 2 < 8) {
        return false;    // Invalid format
      }
-     $calccrc = crc32(substr($data,0,$headerlen)) & 0xffff;
-     $headercrc = unpack("v", substr($data,$headerlen,2));
+     $calccrc = crc32(substr($data, 0, $headerlen)) & 0xffff;
+     $headercrc = unpack("v", substr($data, $headerlen, 2));
      $headercrc = $headercrc[1];
      if ($headercrc != $calccrc) {
        return false;    // Bad header CRC
@@ -1577,18 +1633,18 @@ if (!function_exists('gzdecode')) {
    }
 
    // GZIP FOOTER - These be negative due to PHP's limitations
-   $datacrc = unpack("V",substr($data,-8,4));
+   $datacrc = unpack("V", substr($data, -8, 4));
    $datacrc = $datacrc[1];
-   $isize = unpack("V",substr($data,-4));
+   $isize = unpack("V", substr($data, -4));
    $isize = $isize[1];
 
    // Perform the decompression:
-   $bodylen = $len-$headerlen-8;
+   $bodylen = $len - $headerlen - 8;
    if ($bodylen < 1) {
      // This should never happen - IMPLEMENTATION BUG!
      return null;
    }
-   $body = substr($data,$headerlen,$bodylen);
+   $body = substr($data, $headerlen, $bodylen);
    $data = "";
    if ($bodylen > 0) {
      switch ($method) {
@@ -1613,6 +1669,19 @@ if (!function_exists('gzdecode')) {
      // Bad format!  Length or CRC doesn't match!
      return false;
    }
+
    return $data;
  }
+}
+
+if ( !function_exists('session_status') ){
+    function session_status(){
+        if(!extension_loaded('session')){
+            return 0;
+        }elseif(!session_id()){
+            return 1;
+        }else{
+            return 2;
+        }
+    }
 }
